+++ Source Main 5.2/source/MUMain/stdafx.h
@@ -81,7 +81,7 @@
 
 // Patch
 // WinMain
-#include "Winmain.h"
+#include "MUMain/Winmain.h"
 #include "Defined_Global.h"
 
 // Client
